Slår sammen og flytter på ting
