+++ __tests__/schema/apiPeriodeSchema.test.ts
@@ -1,7 +1,7 @@
-import { PeriodeSchema } from '../../schema/apiPeriodeSchema';
+import { apiPeriodeSchema } from '../../schema/apiPeriodeSchema';
 
 describe('apiPeriodeSchema', () => {
-  it('should not validate wrong PeriodeSchema', () => {
+  it('should not validate wrong apiPeriodeSchema', () => {
     const data = {
       fom: new Date('2021-01-01'),
       tom: new Date('2021-01-02')
@@ -12,12 +12,12 @@ describe('apiPeriodeSchema', () => {
       tom: '2021-01-02'
     };
 
-    const parsed = PeriodeSchema.safeParse(data);
+    const parsed = apiPeriodeSchema.safeParse(data);
     expect(parsed.success).toBe(false);
     expect(parsed.data).not.toEqual(expected);
   });
 
-  it('should validate PeriodeSchema', () => {
+  it('should validate apiPeriodeSchema', () => {
     const data = {
       fom: '2021-01-01',
       tom: '2021-01-02'
@@ -28,7 +28,7 @@ describe('apiPeriodeSchema', () => {
       tom: '2021-01-02'
     };
 
-    const parsed = PeriodeSchema.safeParse(data);
+    const parsed = apiPeriodeSchema.safeParse(data);
     expect(parsed.success).toBe(true);
     expect(parsed.data).toEqual(expected);
   });
