+++ schema/NaturalytelseEnum.ts
@@ -1,0 +1,23 @@
+import { z } from 'zod';
+
+export const NaturalytelseEnum = z.enum([
+  'AKSJERGRUNNFONDSBEVISTILUNDERKURS',
+  'ANNET',
+  'BEDRIFTSBARNEHAGEPLASS',
+  'BESOEKSREISERHJEMMETANNET',
+  'BIL',
+  'BOLIG',
+  'ELEKTRONISKKOMMUNIKASJON',
+  'FRITRANSPORT',
+  'INNBETALINGTILUTENLANDSKPENSJONSORDNING',
+  'KOSTBESPARELSEIHJEMMET',
+  'KOSTDAGER',
+  'KOSTDOEGN',
+  'LOSJI',
+  'OPSJONER',
+  'RENTEFORDELLAAN',
+  'SKATTEPLIKTIGDELFORSIKRINGER',
+  'TILSKUDDBARNEHAGEPLASS',
+  'YRKEBILTJENESTLIGBEHOVKILOMETER',
+  'YRKEBILTJENESTLIGBEHOVLISTEPRIS'
+]);
