+++ schema/begrunnelseRedusertLoennIAgp.ts
@@ -1,0 +1,17 @@
+export const BegrunnelseRedusertLoennIAgp = [
+  'ArbeidOpphoert',
+  'BeskjedGittForSent',
+  'BetvilerArbeidsufoerhet',
+  'FerieEllerAvspasering',
+  'FiskerMedHyre',
+  'FravaerUtenGyldigGrunn',
+  'IkkeFravaer',
+  'IkkeFullStillingsandel',
+  'IkkeLoenn',
+  'LovligFravaer',
+  'ManglerOpptjening',
+  'Permittering',
+  'Saerregler',
+  'StreikEllerLockout',
+  'TidligereVirksomhet'
+] as const;
